requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
schedule==1.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
